Chaged:   * jwt to JWT in function names.   * refactor validateAdmin.   * removed hardcoded jwt parameters
