+++ lib/decorators/extract-jwt.js
@@ -2,7 +2,7 @@
 
 const jsonwebtoken = require('jsonwebtoken');
 
-function getJwt(request) {
+function getJWT(request) {
     if (request.headers.authorization && request.headers.authorization.split(' ')[0] === 'Bearer') {
         return request.headers.authorization.split(' ')[1];
     } else if (request.query && request.query.jwt) {
@@ -11,13 +11,13 @@ function getJwt(request) {
     return undefined;
 }
 
-function decodeJwt(request) {
-    return jsonwebtoken.decode(getJwt(request));
+function decodeJWT(request) {
+    return jsonwebtoken.decode(getJWT(request));
 }
 
-function extractJwt(request, response, next) {
-    request.user = decodeJwt(request);
+function extractJWT(request, response, next) {
+    request.user = decodeJWT(request);
     next();
 }
 
-module.exports = extractJwt;
+module.exports = extractJWT;
