+++ lib/decorators/social-connection.js
@@ -1,12 +1,6 @@
 'use strict';
 
-function validateAdmin(request, response, next) {
-    if (!request.user.profile.roles.includes('admin')) {
-        return response.send(401);
-    }
-
-    return next();
-}
+const validateAdmin = require('./validate-admin');
 
 function decorator(controller) {
     controller.request(validateAdmin);
