+++ lib/models/user.js
@@ -20,7 +20,7 @@ function addConsumer(kongUrl, username, id) {
     });
 }
 
-function addJwtCredentials(kongUrl, username, id, secret) {
+function addJWTCredentials(kongUrl, username, id, secret) {
     return request({
         method: 'POST',
         uri:    `${kongUrl}/consumers/${username}/jwt`,
@@ -70,6 +70,7 @@ const UserSchema = new mongoose.Schema({
 UserSchema.swaggerName = 'User';
 
 UserSchema.pre('save', function(next) {
+    /*eslint no-invalid-this: "off"*/
     this.wasNew = this.isNew;
     if (this.isNew) {
         this.password = sodium.crypto_pwhash_str(new Buffer(this.password),
@@ -84,7 +85,7 @@ UserSchema.post('save', function(doc, next) {
         addConsumer(process.env.EVENTOL_AUTH_KONG_URL, doc.username, doc._id)
             .then((res) => {
                 logger.info('Consumer created: ', res);
-                return addJwtCredentials(process.env.EVENTOL_AUTH_KONG_URL, doc.username,
+                return addJWTCredentials(process.env.EVENTOL_AUTH_KONG_URL, doc.username,
                     doc._id, doc.secret);
             })
             .then((res) => {
