+++ lib/routes/login.js
@@ -1,21 +1,12 @@
 'use strict';
 
-const User = require('../models').User;
-const jwt  = require('jsonwebtoken');
+const User     = require('../models').User;
+const signUser = require('../sign-user');
 
 function logIn(req, res) {
     return User
         .login(req.swagger.params.credentials.value.username, req.swagger.params.credentials.value.password)
-        .then((user) => jwt.sign({
-            profile: {
-                id:    user._id.toString(),
-                roles: user.roles
-            }
-        }, user.secret, {
-            expiresIn: '1h',
-            notBefore: '1s',
-            issuer:    user._id.toString()
-        }))
+        .then(signUser)
         .then((jwt) => res.status(200).send({
             jwt
         }))
