+++ lib/routes/signup.js
@@ -1,20 +1,7 @@
 'use strict';
 
-const User = require('../models').User;
-const jwt  = require('jsonwebtoken');
-
-function signUser(user) {
-    return jwt.sign({
-        profile: {
-            id:    user._id.toString(),
-            roles: user.roles
-        }
-    }, user.secret, {
-        expiresIn: '1h',
-        notBefore: '1s',
-        issuer:    user._id.toString()
-    });
-}
+const User     = require('../models').User;
+const signUser = require('../sign-user');
 
 function signUp(req, res) {
     return User
